+++ src/layouts/InAppLayout.tsx
@@ -1,5 +1,5 @@
 // Core
-import React, { useEffect, useState } from "react";
+import React, { useState } from "react";
 
 // Components
 import Sidebar from "@/components/sidebar/app-sidebar";
